Added a table and appending items to it
I've added a table to hold the steps and results of the algorithm. You can now select a number between 1 and 15000 and when you click the test button, the number will be put through the test and the results are then appended to the table. To ensure that the table is cleared between test, all children elements of the table are removed when the button is clicked.